+++ src/main/java/gov/usgs/owi/nldi/services/Parameters.java
@@ -29,6 +29,7 @@ public class Parameters {
 	public static final String LEGACY = "legacy";
 	public static final String LATITUDE = "lat";
 	public static final String LONGITUDE = "lon";
+	public static final String SIMPLIFIED = "simplified";
 
 	// Validates that the "coords" parameter is in the form POINT(-89.35 43.0864))
 	public static final String POINT_VALIDATION_MESSAGE = "coords must be specified as a point with longitude and latitude, i.e. POINT(-89.35 43.0864)";