+++ src/test/java/gov/usgs/owi/nldi/controllers/BaseControllerTest.java
@@ -114,7 +114,16 @@ public class BaseControllerTest {
 
 	@Test
 	public void streamBasinTest() throws Exception {
-		controller.streamBasin(response, "123");
+		controller.streamBasin(response, "123", true);
+		verify(streamingDao).stream(anyString(), anyMap(), any(ResultHandler.class));
+		verify(navigation, never()).navigate(anyMap());
+		verify(navigation, never()).interpretResult(anyMap(), any(HttpServletResponse.class));
+		assertEquals(HttpStatus.OK.value(), response.getStatus());
+	}
+
+	@Test
+	public void streamBasinNonSimplifiedTest() throws Exception {
+		controller.streamBasin(response, "123", false);
 		verify(streamingDao).stream(anyString(), anyMap(), any(ResultHandler.class));
 		verify(navigation, never()).navigate(anyMap());
 		verify(navigation, never()).interpretResult(anyMap(), any(HttpServletResponse.class));