+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/ACWI-SSWD/nldi-services/compare/nldi-services-1.3.0...master)
+* Add `?simplified=false` parameter to retrieve full resolution basin boundary
 
 ## [1.3.0](https://github.com/ACWI-SSWD/nldi-services/compare/nldi-services-1.2.0...nldi-services-1.3.0)
 ### Changed